Move RunRecorder implementations to package
